Modify AWS CloudWatch agent to start after cloud-init
We sometimes use cloud-init to set the hostname to the correct value,
and introducing this dependency ensures that this is allowed to happen
before the CloudWatch Agent starts up.

Having a service run after cloud-init.target is trickier than one
might expect because of a dependency cycle that is induced because the
default of DefaultDependencies=yes causes the default target
(usually multi-user.target) to depend on the service.  This is
described in great detail here:
https://serverfault.com/a/973985

Note also that we are using a drop-in file so that we can leave the
real unit file untouched and need only specify the unit file
attributes that we want to change.