Remove recurse attribute for ansible.builtin.file task
This attribute is not necessary, and it causes an idempotence failure
now that we specify the mode for the override.conf file in the
following task.  The reason for this is that we specify the mode of
the override.conf file as 0644, but the recurse attribute causes the
mode specified in the directory creation task (0755) to be applied to
the contents of the directory.  These contents include the
override.conf file.

Now that the recurse attribute is not present, we must add the state
attribute with a value of directory.